refactor: abstract validator and fix imports
