+++ src/utils/validator/isTextValid.go
